requirements.txt
# Sanskrit RAG — CPU-only requirements
# Install: pip install -r requirements.txt

sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
transformers>=4.40.0
torch>=2.2.0
numpy>=1.26.0
streamlit>=1.35.0
